+++ lib/doggo.ex
@@ -191,6 +191,155 @@ defmodule Doggo do
   end
 
   @doc """
+  Renders an alert dialog that requires the immediate attention and response of
+  the user.
+
+  This component is meant for situations where critical information must be
+  conveyed, and an explicit response is required from the user. It is typically
+  used for confirmation dialogs, warning messages, error notifications, and
+  other scenarios where an immediate decision is necessary.
+
+  For non-critical dialogs, such as those containing forms or additional
+  information, use `Doggo.modal/1` instead.
+
+  ## Usage
+
+  ```heex
+  <Doggo.alert_dialog id="end-session-modal">
+    <:title>End Training Session Early?</:title>
+    <p>
+      Are you sure you want to end the current training session with Bella?
+      She's making great progress today!
+    </p>
+    <:footer>
+      <Doggo.button phx-click="end-session">
+        Yes, end session
+      </Doggo.button>
+      <Doggo.button phx-click={JS.exec("data-cancel", to: "#end-session-modal")}>
+        No, continue training
+      </Doggo.button>
+    </:footer>
+  </Doggo.alert_dialog>
+  ```
+
+  To open the dialog, use the `show_modal/1` function.
+
+  ```heex
+  <Doggo.button
+    phx-click={Doggo.show_modal("end-session-modal")}
+    aria-haspopup="dialog"
+  >
+    show
+  </Doggo.button>
+  ```
+
+  ## CSS
+
+  To hide the modal when the `open` attribute is not set, use the following CSS
+  styles:
+
+  ```css
+  dialog.alert-dialog:not([open]),
+  dialog.alert-dialog[open="false"] {
+    display: none;
+  }
+  ```
+
+  ## Semantics
+
+  While the `showModal()` JavaScript function is typically recommended for
+  managing modal dialog semantics, this component utilizes the `open` attribute
+  to control visibility. This approach is chosen to eliminate the need for
+  library consumers to add additional JavaScript code. To ensure proper
+  modal semantics, the `aria-modal` attribute is added to the dialog element.
+  """
+  @doc type: :component
+
+  attr :id, :string, required: true
+  attr :open, :boolean, default: false, doc: "Initializes the dialog as open."
+
+  attr :on_cancel, JS,
+    default: %JS{},
+    doc: """
+    An additional `Phoenix.LiveView.JS` command to execute when the dialog
+    is canceled. This command is executed in addition to closing the dialog. If
+    you only want the dialog to be closed, you don't have to set this attribute.
+    """
+
+  attr :dismissable, :boolean,
+    default: false,
+    doc: """
+    When set to `true`, the dialog can be dismissed by clicking a close button
+    or by pressing the escape key.
+    """
+
+  attr :close_label, :string,
+    default: "Close",
+    doc: "Aria label for the close button."
+
+  slot :title, required: true
+  slot :inner_block, required: true, doc: "The modal body."
+
+  slot :close,
+    doc: "The content for the 'close' link. Defaults to the word 'close'."
+
+  slot :footer
+
+  attr :class, :any,
+    default: [],
+    doc: "Additional CSS classes. Can be a string or a list of strings."
+
+  attr :rest, :global, doc: "Any additional HTML attributes."
+
+  def alert_dialog(assigns) do
+    ~H"""
+    <dialog
+      id={@id}
+      role="alertdialog"
+      class={["alert-dialog" | List.wrap(@class)]}
+      aria-modal={(@open && "true") || "false"}
+      aria-labelledby={"#{@id}-title"}
+      aria-describedby={"#{@id}-content"}
+      open={@open}
+      phx-mounted={@open && show_modal(@id)}
+      phx-remove={hide_modal(@id)}
+      data-cancel={JS.exec(@on_cancel, "phx-remove")}
+      {@rest}
+    >
+      <.focus_wrap
+        id={"#{@id}-container"}
+        class="alert-dialog-container"
+        phx-window-keydown={@dismissable && JS.exec("data-cancel", to: "##{@id}")}
+        phx-key={@dismissable && "escape"}
+        phx-click-away={@dismissable && JS.exec("data-cancel", to: "##{@id}")}
+      >
+        <section>
+          <header>
+            <button
+              :if={@dismissable}
+              href="#"
+              class="alert-dialog-close"
+              aria-label={@close_label}
+              phx-click={JS.exec("data-cancel", to: "##{@id}")}
+            >
+              <%= render_slot(@close) %>
+              <span :if={@close == []}>close</span>
+            </button>
+            <h2 id={"#{@id}-title"}><%= render_slot(@title) %></h2>
+          </header>
+          <div id={"#{@id}-content"} class="alert-dialog-content">
+            <%= render_slot(@inner_block) %>
+          </div>
+          <footer :if={@footer != []}>
+            <%= render_slot(@footer) %>
+          </footer>
+        </section>
+      </.focus_wrap>
+    </dialog>
+    """
+  end
+
+  @doc """
   The app bar is typically located at the top of the interface and provides
   access to key features and navigation options.
 
@@ -2987,7 +3136,11 @@ defmodule Doggo do
   end
 
   @doc """
-  Renders a modal.
+  Renders a modal dialog for content such as forms and informational panels.
+
+  This component is appropriate for non-critical interactions. For dialogs
+  requiring immediate user response, such as confirmations or warnings, use
+  `Doggo.alert_dialog/1` instead.
 
   ## Usage
 
@@ -3087,7 +3240,21 @@ defmodule Doggo do
 
   attr :id, :string, required: true
   attr :open, :boolean, default: false, doc: "Initializes the modal as open."
-  attr :on_cancel, JS, default: %JS{}
+
+  attr :on_cancel, JS,
+    default: %JS{},
+    doc: """
+    An additional `Phoenix.LiveView.JS` command to execute when the dialog
+    is canceled. This command is executed in addition to closing the dialog. If
+    you only want the dialog to be closed, you don't have to set this attribute.
+    """
+
+  attr :dismissable, :boolean,
+    default: true,
+    doc: """
+    When set to `true`, the dialog can be dismissed by clicking a close button
+    or by pressing the escape key.
+    """
 
   attr :close_label, :string,
     default: "Close",
@@ -3123,13 +3290,14 @@ defmodule Doggo do
       <.focus_wrap
         id={"#{@id}-container"}
         class="modal-container"
-        phx-window-keydown={JS.exec("data-cancel", to: "##{@id}")}
-        phx-key="escape"
-        phx-click-away={JS.exec("data-cancel", to: "##{@id}")}
+        phx-window-keydown={@dismissable && JS.exec("data-cancel", to: "##{@id}")}
+        phx-key={@dismissable && "escape"}
+        phx-click-away={@dismissable && JS.exec("data-cancel", to: "##{@id}")}
       >
-        <article>
+        <section>
           <header>
-            <.link
+            <button
+              :if={@dismissable}
               href="#"
               class="modal-close"
               aria-label={@close_label}
@@ -3137,7 +3305,7 @@ defmodule Doggo do
             >
               <%= render_slot(@close) %>
               <span :if={@close == []}>close</span>
-            </.link>
+            </button>
             <h2 id={"#{@id}-title"}><%= render_slot(@title) %></h2>
           </header>
           <div id={"#{@id}-content"} class="modal-content">
@@ -3146,7 +3314,7 @@ defmodule Doggo do
           <footer :if={@footer != []}>
             <%= render_slot(@footer) %>
           </footer>
-        </article>
+        </section>
       </.focus_wrap>
     </dialog>
     """