+++ test/doggo_test.exs
@@ -342,6 +342,148 @@ defmodule DoggoTest do
     end
   end
 
+  describe "alert_dialog/1" do
+    test "default" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" on_cancel={JS.push("cancel")}>
+          <:title>Edit dog</:title>
+          dog-form
+          <:footer>paw</:footer>
+        </Doggo.alert_dialog>
+        """)
+
+      dialog = find_one(html, "dialog:root")
+      assert attribute(dialog, "role") == "alertdialog"
+      assert attribute(dialog, "id") == "pet-alert"
+      assert attribute(dialog, "class") == "alert-dialog"
+      assert attribute(dialog, "aria-modal") == "false"
+      assert attribute(dialog, "open") == nil
+      assert attribute(dialog, "phx-mounted") == nil
+
+      assert Floki.find(html, ".alert-dialog-close") == []
+
+      h2 = find_one(html, ":root > div > section > header > h2")
+      assert text(h2) == "Edit dog"
+
+      assert text(html, "section > .alert-dialog-content") == "dog-form"
+      assert text(html, "section > footer") == "paw"
+    end
+
+    test "dismissable" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" on_cancel={JS.push("cancel")} dismissable>
+          <:title>Edit dog</:title>
+          dog-form
+          <:footer>paw</:footer>
+        </Doggo.alert_dialog>
+        """)
+
+      a =
+        find_one(
+          html,
+          ":root > div > section > header > button.alert-dialog-close"
+        )
+
+      assert attribute(a, "aria-label") == "Close"
+      assert text(a, "span") == "close"
+    end
+
+    test "opened" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" open>
+          <:title>Edit dog</:title>
+          dog-form
+        </Doggo.alert_dialog>
+        """)
+
+      dialog = find_one(html, "dialog:root")
+      assert attribute(dialog, "aria-modal") == "true"
+      assert attribute(dialog, "open") == "open"
+    end
+
+    test "with close slot" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" dismissable>
+          <:title>Edit dog</:title>
+          dog-form
+          <:close>X</:close>
+        </Doggo.alert_dialog>
+        """)
+
+      assert text(html, "button.alert-dialog-close") == "X"
+    end
+
+    test "with close label" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" close_label="Cancel" dismissable>
+          <:title>Edit dog</:title>
+          dog-form
+        </Doggo.alert_dialog>
+        """)
+
+      assert attribute(html, "button.alert-dialog-close", "aria-label") ==
+               "Cancel"
+    end
+
+    test "with additional class as string" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" class="is-narrow">
+          <:title>Edit dog</:title>
+          dog-form
+        </Doggo.alert_dialog>
+        """)
+
+      assert attribute(html, ":root", "class") == "alert-dialog is-narrow"
+    end
+
+    test "with additional classes as list" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" class={["is-narrow", "is-dark"]}>
+          <:title>Edit dog</:title>
+          dog-form
+        </Doggo.alert_dialog>
+        """)
+
+      assert attribute(html, ":root", "class") ==
+               "alert-dialog is-narrow is-dark"
+    end
+
+    test "with global attribute" do
+      assigns = %{}
+
+      html =
+        parse_heex(~H"""
+        <Doggo.alert_dialog id="pet-alert" data-test="hello">
+          <:title>Edit dog</:title>
+          dog-form
+        </Doggo.alert_dialog>
+        """)
+
+      assert attribute(html, ":root", "data-test") == "hello"
+    end
+  end
+
   describe "app_bar/1" do
     test "default" do
       assigns = %{}